Done restructuring login
It now follows the same system as the home page
